+++ src/Oriflame.PolicyBuilder.Policies/Builders/Fluent/Sections/IOutboundSectionPolicyBuilder.cs
@@ -1,5 +1,7 @@
 ﻿using System;
+using System.Collections.Generic;
 using Oriflame.PolicyBuilder.Policies.Builders.Fluent.Actions;
+using Oriflame.PolicyBuilder.Policies.Builders.Fluent.Attributes;
 using Oriflame.PolicyBuilder.Policies.DynamicProperties;
 
 namespace Oriflame.PolicyBuilder.Policies.Builders.Fluent.Sections
@@ -13,9 +15,9 @@ namespace Oriflame.PolicyBuilder.Policies.Builders.Fluent.Sections
         IOutboundSectionPolicyBuilder SetBody(string content);
 
         /// <see href="https://docs.microsoft.com/en-us/azure/api-management/api-management-caching-policies#StoreToCache"/>
-        IOutboundSectionPolicyBuilder CacheStore(TimeSpan cacheDuration);
+        IOutboundSectionPolicyBuilder CacheStore(TimeSpan cacheDuration, Func<ICacheStoreOptionalAttributesBuilder, IDictionary<string, string>> cacheStoreOptionalAttributesBuilder = null);
         
         /// <see href="https://docs.microsoft.com/en-us/azure/api-management/api-management-caching-policies#StoreToCache"/>
-        IOutboundSectionPolicyBuilder CacheStore(string cacheDuration);
+        IOutboundSectionPolicyBuilder CacheStore(string cacheDuration, Func<ICacheStoreOptionalAttributesBuilder, IDictionary<string, string>> cacheStoreOptionalAttributesBuilder = null);
     }
 }